feat: wip switch to free type text pt6

+++ PHEngine/PHEngine/src/Core/GameCore/Components/UiComponents/UiComponent.cpp
@@ -31,7 +31,7 @@ void UiComponent::CollectDataForSerialization(SerializeDataContainer& dataContai
 
 int32_t UiComponent::CreateTextField(
     const std::string& fontName,
-    const float fontSize,
+    const int32_t fontSize,
     const std::string& text,
     const glm::vec3& color,
     const glm::vec2& position,
@@ -60,7 +60,7 @@ int32_t UiComponent::CreateTextField(
 
 int32_t UiComponent::CreateEmptyTextField(
     const std::string& fontName,
-    const float fontSize,
+    const int32_t fontSize,
     const glm::vec3& color,
     const bool receiveUpdateOnTextScreenSpaceSizeChanged,
     const float lineMaxSize,
