+++ PHEngine/PHEngine/src/Core/GameCore/GUI/HudText/HudTextField.cpp
@@ -7,7 +7,7 @@
 namespace EngineCore {
 HudTextField::HudTextField(
     const std::string& fontName,
-    const float fontSize,
+    const int32_t fontSize,
     const std::string& text,
     const glm::vec3& color,
     const glm::vec2& position,
@@ -30,7 +30,7 @@ HudTextField::HudTextField(
 
 HudTextField::HudTextField(
     const std::string& fontName,
-    const float fontSize,
+    const int32_t fontSize,
     const glm::vec3& color,
     const float lineMaxSize,
     const int32_t numberOfLines,
@@ -68,7 +68,7 @@ std::string HudTextField::GetFontName() const
     return mFontName;
 }
 
-float HudTextField::GetFontSize() const
+int32_t HudTextField::GetFontSize() const
 {
     return mFontSize;
 }
