+++ PHEngine/PHEngine/src/Core/GameCore/ScriptingCore/LuaProxies/UiTextBlockLuaProxy.h
@@ -21,7 +21,7 @@ protected:
 
     const std::string mFontName;
 
-    float mFontSize;
+    int32_t mFontSize;
 
     float mTextLineWidth;
 
@@ -40,7 +40,7 @@ public:
 
     void SetOpacity_FromGameThread(const float opacity);
 
-    void SetFontSize_FromGameThread(const float fontSize);
+    void SetFontSize_FromGameThread(const int32_t fontSize);
 
     void SetTextLineWidth_FromGameThread(const float textLineWidth);
